// src/stores/theme/types/paginationTypes.ts
// Pagination component type definition

export interface PaginationConfig {
  background: string;
  borderColor: string;
  textColor: string;
  buttonBackground: string;
  buttonTextColor: string;
  buttonHoverBackground: string;
  buttonHoverTextColor: string;
  buttonActiveBackground: string;
  buttonActiveTextColor: string;
  buttonBorderRadius: string;
  navTextColor: string;
  navDisabledColor: string;
  selectBackground: string;
  selectBorderColor: string;
  selectTextColor: string;
  infoTextColor: string;
}